Types

Index

Muon.AnnData

Muon.MuData

All types

Muon.AnnData — Type

An annotated data object that stores data matrices with associated metadata.

Constructor

AnnData(;
    X::AbstractMatrix{<:Number},
    obs::Union{DataFrame, Nothing}=nothing,
    obs_names::Union{AbstractVector{<:AbstractString}, Nothing}=nothing,
    var::Union{DataFrame, Nothing}=nothing,
    var_names::Union{AbstractVector{<:AbstractString}, Nothing}=nothing,
    obsm::Union{AbstractDict{<:AbstractString, <:Union{<:AbstractArray{<:Number}, DataFrame}}, Nothing}=nothing,
    varm::Union{AbstractDict{<:AbstractString, <:Union{<:AbstractArray{<:Number}, DataFrame}}, Nothing}=nothing,
    obsp::Union{AbstractDict{<:AbstractString, <:AbstractMatrix{<:Number}}, Nothing}=nothing,
    varp::Union{AbstractDict{<:AbstractString, <:AbstractMatrix{<:Number}}, Nothing}=nothing,
    layers::Union{AbstractDict{<:AbstractString, <:AbstractMatrix{<:Number}}, Nothing}=nothing,
    uns::Union{AbstractDict{<:AbstractString, <:Any}, Nothing}=nothing,
)

Keyword arguments / fields of the object

X: An observations × variables matrix.

obs: A DataFrame with observation-level metadata.

obs_names: A vector of observation names (identifiers).

var: A DataFrame with variable-level metadata.

var_names: A vector of ariable names (identifiers).

obsm: Dictionary with observation-level metadata.

varm: Dictionary of observation-level metadata.

obsp: Dictionary of pairwise observation-level metadata. Each element of obsp is a square matrix.

varp: Dictionary of pairwise variable-level metadata. Each element of varp is a square matrix.

layers: Dictionary of additional observations × variables matrices, e.g. for different processing/normalization steps.

uns: Dictionary with unstructured metadata.

Muon.MuData — Type

A multimodal data object that stores collections of AnnData objects.

Constructor

MuData(;
    mod::Union{AbstractDict{<:AbstractString, AnnData}, Nothing}=nothing,
    obs::Union{DataFrame, Nothing}=nothing,
    obs_names::Union{AbstractVector{<:AbstractString}, Nothing}=nothing,
    var::Union{DataFrame, Nothing}=nothing,
    var_names::Union{AbstractVector{<:AbstractString}, Nothing}=nothing,
    obsm::Union{AbstractDict{<:AbstractString, <:Union{<:AbstractArray{<:Number}, DataFrame}}, Nothing}=nothing,
    varm::Union{AbstractDict{<:AbstractString, <:Union{<:AbstractArray{<:Number}, DataFrame}}, Nothing}=nothing,
    obsp::Union{AbstractDict{<:AbstractString, <:AbstractMatrix{<:Number}}, Nothing}=nothing,
    varp::Union{AbstractDict{<:AbstractString, <:AbstractMatrix{<:Number}}, Nothing}=nothing,
    obsmap::Union{AbstractDict{<:AbstractString, <:AbstractVector{<:Integer}}, Nothing}=nothing,
    varmap::Union{AbstractDict{<:AbstractString, <:AbstractVector{<:Integer}}, Nothing}=nothing,
    uns::Union{AbstractDict{<:AbstractString, <:Any}, Nothing}=nothing,
    axis::Union{Integer, Nothing}=nothing,
)

Keyword arguments / fields of the object

mod: A dictionary of AnnData objects.

obs: A DataFrame with observation-level metadata.

obs_names: A vector of observation names (identifiers). The names correspond to obs_names of the AnnData objects.

var: A DataFrame with variable-level metadata

var_names: A vector of variable names (identifiers). The names correspond to var_names of the AnnData objects.

obsm: Dictionary of observation-level metadata.

varm: Dictionary of variable-level metadata.

obsp: Dictionary of pairwise observation-level metadata. Each element of obsp is a square matrix.

varp: Dictionary of pairwise variable-level metadata. Each element of varp is a square matrix.

obsmap: Dictionary with one integer vector for each modality. The vectors contain for each observation in the MuData object the index of corresponding observation in the respective AnnData object, or 0 if the observation is not present. This argument to the constructor should generally not be required unless you know what you're doing.

varmap: Dictionary with one integer vector for each modality. The vectors contain for each variable in the MuData object the index of the corresponding variable in the respective AnnData object, or 0 if the observation is not present. This argument to the constructor should generally not be required unless you know what you're doing.

uns: Dictionary with unstructured metadata.

axis: Axis of the object. axis=1 (default) indicates sharing of observations, each AnnData object is treated as a separate modality with no overlapping variables. axis=2 indicates sharing of variables, each AnnData is treated as a dataset with no overlapping observations. axis=0 indicates sharing of both observations and variables, useful when the different AnnData objects are subsets of the same dataset.